subprojects/flare-md-gateway/gateway/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
readability-lxml==0.8.1
markdownify==0.13.1
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.1.0
python-multipart==0.0.9
pydantic==2.8.2
